updated values retrieving system
- Renamed `ValueField` into `Written`.
- Removed `SelectMenu`, replaced by `Selected`.
These changes mark values retrieved as 2 types:
written values and selected values.
- Implemented `merge` functions for `Format`s.
Now inheriting a format from the `Values` on a field is smarter,
saving custom formatting specs.

- `ValueMenu` renamed to `Values`.
`Values` now serves as a `MenuStream` holder
and to inherit a global `Format` on the next values.
The struct does not hold fields anymore.